refactor: add encrypter factory
