+++ pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   shared_preferences: 2.2.2
   encrypt: 5.0.3
   crypto: 3.0.3
+
   flutter:
     sdk: flutter
 
@@ -20,6 +21,7 @@ dev_dependencies:
   flutter_lints: ">=2.0.0 <4.0.0"
   mocktail: 1.0.3
   melos: 5.3.0
+
   flutter_test:
     sdk: flutter
